+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authz/requests/TupleCreateRequest.java
@@ -2,9 +2,11 @@ package cloud.pangeacyber.pangea.authz.requests;
 
 import cloud.pangeacyber.pangea.BaseRequest;
 import cloud.pangeacyber.pangea.authz.models.Tuple;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class TupleCreateRequest extends BaseRequest {
 
+	@JsonProperty("tuples")
 	private Tuple[] tuples;
 
 	private TupleCreateRequest(Builder builder) {