+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- `attributes` field in `/list-resources` and `/list-subjects` endpoint
+
 ### Fixed
 
 - Value of `TransferMethod.SOURCE_URL`.